backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-generativeai
python-dotenv
SQLAlchemy
Pillow
requests
